Se ajustan los repositorios

+++ logistica_transporte_backend/src/main/java/com/logistica/transporte/repository/BodegaRepository.java
@@ -3,7 +3,6 @@ package com.logistica.transporte.repository;
 import com.logistica.transporte.entity.Bodega;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 
-public interface BodegaRepository extends JpaRepository<Bodega, BigInteger> {
+public interface BodegaRepository extends JpaRepository<Bodega, Long> {
 }
