+++ logistica_transporte_backend/src/main/java/com/logistica/transporte/repository/LogisticaMaritimaRepository.java
@@ -3,7 +3,6 @@ package com.logistica.transporte.repository;
 import com.logistica.transporte.entity.LogisticaMaritima;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 
-public interface LogisticaMaritimaRepository extends JpaRepository<LogisticaMaritima, BigInteger> {
+public interface LogisticaMaritimaRepository extends JpaRepository<LogisticaMaritima, Long> {
 }
