+++ logistica_transporte_backend/src/main/java/com/logistica/transporte/repository/LogisticaTerrestreRepository.java
@@ -3,7 +3,6 @@ package com.logistica.transporte.repository;
 import com.logistica.transporte.entity.LogisticaTerrestre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 
-public interface LogisticaTerrestreRepository extends JpaRepository<LogisticaTerrestre, BigInteger> {
+public interface LogisticaTerrestreRepository extends JpaRepository<LogisticaTerrestre, Long> {
 }
