+++ logistica_transporte_backend/src/main/java/com/logistica/transporte/repository/ProductoRepository.java
@@ -3,7 +3,6 @@ package com.logistica.transporte.repository;
 import com.logistica.transporte.entity.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 
-public interface ProductoRepository extends JpaRepository<Producto, BigInteger> {
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
 }
