+++ logistica_transporte_backend/src/main/java/com/logistica/transporte/repository/PuertoMaritimoRepository.java
@@ -3,7 +3,6 @@ package com.logistica.transporte.repository;
 import com.logistica.transporte.entity.PuertoMaritimo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 
-public interface PuertoMaritimoRepository extends JpaRepository<PuertoMaritimo, BigInteger> {
+public interface PuertoMaritimoRepository extends JpaRepository<PuertoMaritimo, Long> {
 }
